Mushrooms Classification using the most -ve Correlated Features w.r.t Target
Did a test on whether -ve correlated features are good for Model Classification or not. Found out they also contribute the same towards the target classification. In fact, i was able to receive 100% accuracy ( train, test & CV) when i included most of the -ve features along with 2-3 +Vely correlated features.